Add a Title model
Added a Title model to hold information about a run of issues. Although issues
are contained within a title, they won't be held with the title model as this
would mean all issues would have to be loaded into the model on load.

Instead the Title has summary information about the titles (start and end year
as well as issue count).