Add an In Memory Connection
Added a Connection for repositories to use to store transitory data. As the
data is stored in an array, any data stored will not last past the completion
of the script. This is useful for making unit test repositories or for storing
data that doesn't need to be accessed again.