Added InMemory implemenation of an Issue repository
Moved code from the Title InMemory repository to a base class so it can reused
by the new Issue repository. This repository is an implementation of the Issue
Repository interface for retrieving details about an issue. It works in a
similar way to the Title interface and they are very interrelated due to how
Issue information derives Title summary